feat: add depth limit
