+++ src/routes/graphql/index.ts
@@ -1,7 +1,39 @@
 import { FastifyPluginAsyncTypebox } from '@fastify/type-provider-typebox';
 import { createGqlResponseSchema, gqlResponseSchema } from './schemas.js';
-import { graphql } from 'graphql';
-import rootSchema from './schemas/schema.js';
+import { graphql, parse, validate, GraphQLObjectType, GraphQLSchema } from 'graphql';
+import depthLimit from 'graphql-depth-limit';
+import { userQueries }  from './schemas/user/queries.js';
+import { profileQueries }  from './schemas/profile/queries.js';
+import { postQueries }  from './schemas/post/queries.js';
+import { memberTypeQueries }  from './schemas/memberType/queries.js';
+import { postMutations } from './schemas/post/mutation.js';
+import { profileMutations } from './schemas/profile/mutation.js';
+import { userMutations } from './schemas/user/mutation.js';
+
+const queryFields = () => ({
+  ...userQueries,
+  ...profileQueries,
+  ...postQueries,
+  ...memberTypeQueries,
+})
+
+const mutationFields = () => ({
+  ...postMutations,
+  ...profileMutations,
+  ...userMutations
+})
+
+const schema = new GraphQLSchema({
+  query: new GraphQLObjectType({
+    name: 'QueryType',
+    fields: queryFields
+  }),
+  mutation: new GraphQLObjectType({
+    name: 'MutationType',
+    fields: mutationFields
+  }),
+});
+
 
 const plugin: FastifyPluginAsyncTypebox = async (fastify) => {
   fastify.route({
@@ -14,12 +46,16 @@ const plugin: FastifyPluginAsyncTypebox = async (fastify) => {
       },
     },
     async handler(req) {
-      return await graphql({
-        schema: rootSchema,
+      const validateErrors = validate(schema, parse(req.body.query), [
+        depthLimit(5),
+      ]);
+
+      return validateErrors && validateErrors.length != 0 ? { data: '', errors: validateErrors } : await graphql({
+        schema: schema,
         source: req.body.query,
         variableValues: req.body.variables,
         contextValue: { prismaClient: fastify.prisma },
-      });
+      })
     },
   });
 };
